This adds the option to override the behave options in config.ini.
It has one really odd snag which I have not found the reason for. If the option "--format <format>" is omitted, behave crashes:

  File "/usr/local/lib/python2.7/dist-packages/behave/formatter/formatters.py", line 45, in get_formatter
    for i, name in enumerate(config.format):

Behave itself has no issues with just --tags=foo, but apitest works only because --format pretty is in the behave part of config.ini.. But since that's default, using this feature looks buggy.